Updated README to use the right case for the `set_property` function
The examples setting dimensions mentioned to use `setProperty` when dealing with high cardinality metrics. This should instead refer to `set_property`, which is the Python method and the one referred to throughout the rest of the README. The current casing might lead to confusion and make users believe there is a separate `setProperty` function.